+++ jd_price.js
@@ -18,171 +18,287 @@ cron "39 0,18 * * *" script-path=https://raw.githubusercontent.com/KingRan/JDJB/
 ============小火箭=========
 京东保价 = type=cron,script-path=https://raw.githubusercontent.com/KingRan/JDJB/main/jd_price.js, cronexpr="39 0,18 * * *", timeout=3600, enable=true
  */
- 
-let common = require("./function/common");
-let jsdom = require("jsdom");
-let $ = new common.env('京东保价');
-let min = 1,
-    help = $.config[$.filename(__filename)] || Math.min(min, $.config.JdMain) || min;
-$.setOptions({
-    headers: {
-        'content-type': 'application/json',
-        'user-agent': 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10.15; rv:91.0) Gecko/20100101 Firefox/91.0',
-        'referer': 'https://msitepp-fm.jd.com/rest/priceprophone/priceProPhoneMenu',
+const $ = new Env('京东保价');
+const notify = $.isNode() ? require('./sendNotify') : '';
+//Node.js用户请在jdCookie.js处填写京东ck;
+const jdCookieNode = $.isNode() ? require('./jdCookie.js') : '';
+const jsdom = $.isNode() ? require('jsdom') : '';
+//IOS等用户直接用NobyDa的jd cookie
+let cookiesArr = [], cookie = '', message, allMessage = '';
+if ($.isNode()) {
+    Object.keys(jdCookieNode).forEach((item) => {
+        cookiesArr.push(jdCookieNode[item])
+    })
+    if (process.env.JD_DEBUG && process.env.JD_DEBUG === 'false') console.log = () => {};
+} else {
+    cookiesArr = [$.getdata('CookieJD'), $.getdata('CookieJD2'), ...jsonParse($.getdata('CookiesJD') || "[]").map(item => item.cookie)].filter(item => !!item);
+}
+
+const JD_API_HOST = 'https://api.m.jd.com/';
+!(async () => {
+    if (!cookiesArr[0]) {
+        $.msg($.name, '【提示】请先获取京东账号一cookie\n直接使用NobyDa的京东签到获取', 'https://bean.m.jd.com/bean/signIndex.action', {"open-url": "https://bean.m.jd.com/bean/signIndex.action"});
+        return;
     }
-});
-$.readme = `
-48 */8 * * * task ${$.runfile}
-exprot ${$.runfile}=1  #输出购买订单保价内容,没什么用
-`
-eval(common.eval.mainEval($));
-async function prepare() {}
-async function main(id) {
-    try {
-        await jstoken()
-        // 一键保价
-        p = {
-            'url': `https://api.m.jd.com/api?appid=siteppM&functionId=siteppM_skuOnceApply&forcebot=&t=${$.timestamp}`,
-            "form": {
-                "body": JSON.stringify({
-                    sid: '',
-                    type: 3,
-                    forcebot: '',
-                    token: $.token,
-                    feSt: 's'
-                })
-            }
-        };
-        h = await $.curl(p)
-        console.log(h)
-        console.log("等待20s,获取保价订单")
-        await $.wait(20000)
-        // 获取保价信息
-        let p2 = {
-            'url': `https://api.m.jd.com/api?appid=siteppM&functionId=siteppM_appliedSuccAmount&forcebot=&t=${$.timestamp}`,
-            // 'form': {
-            //     'body': `"{\"sid\":\"\",\"type\":\"3\",\"forcebot\":\"\"}"`
-            // }
-            'form': 'body={"sid":"","type":"3","forcebot":"","num":15}'
-        }
-        await $.curl(p2)
-        if ($.source.flag) {
-            text = `本次保价金额: ${$.source.succAmount}`
-        } else {
-            text = "本次无保价订单"
-        }
-        console.log(text)
-        $.notice(text)
-        if ($.config[$.runfile]) {
-            // 单个商品检测,没什么用处
-            console.log("\n手动保价前25个订单")
-            html = ''
-            for (let i = 1; i < 6; i++) {
-                await jstoken()
-                p3 = {
-                    'url': `https://api.m.jd.com/api?appid=siteppM&functionId=siteppM_priceskusPull&forcebot=&t=${$.timestamp}`,
-                    'form': {
-                        'body': JSON.stringify({
-                            "page": i,
-                            "pageSize": 5,
-                            "keyWords": "",
-                            "sid": "",
-                            "type": "3",
-                            "forcebot": "",
-                            "token": $.token,
-                            "feSt": "s"
-                        })
-                    }
+    await jstoken();
+    for (let i = 0; i < cookiesArr.length; i++) {
+        if (cookiesArr[i]) {
+            cookie = cookiesArr[i];
+            $.UserName = decodeURIComponent(cookie.match(/pt_pin=([^; ]+)(?=;?)/) && cookie.match(/pt_pin=([^; ]+)(?=;?)/)[1])
+            $.index = i + 1;
+            $.isLogin = true;
+            $.nickName = '';
+            $.token = ''
+            message = '';
+            await TotalBean();
+            console.log(`\n******开始【京东账号${$.index}】${$.nickName || $.UserName}*********\n`);
+            if (!$.isLogin) {
+                $.msg($.name, `【提示】cookie已失效`, `京东账号${$.index} ${$.nickName || $.UserName}\n请重新登录获取\nhttps://bean.m.jd.com/bean/signIndex.action`, {"open-url": "https://bean.m.jd.com/bean/signIndex.action"});
+
+                if ($.isNode()) {
+                    await notify.sendNotify(`${$.name}cookie已失效 - ${$.UserName}`, `京东账号${$.index} ${$.UserName}\n请重新登录获取cookie`);
                 }
-                html += await $.curl(p3)
+                continue
             }
-            amount = $.matchall(/class="name"\>\s*([^\<]+).*?orderId="(\d+)"\s*skuId="(\d+)"/g, html.replace(/\n/g, ''))
-            for (let i of amount) {
-                // 获取有无申请按钮
-                p4 = {
-                    'url': `https://api.m.jd.com/api?appid=siteppM&functionId=siteppM_skuProResultPin&forcebot=&t=${$.timestamp}`,
-                    'form': {
-                        'body': JSON.stringify({
-                            "orderId": i[1],
-                            "skuId": i[2],
-                            "sequence": "1",
-                            "sid": "",
-                            "type": "3",
-                            "forcebot": ""
-                        })
-                    }
-                }
-                h = await $.curl(p4)
-                if (h.includes("hidden")) {
-                    console.log(`商品: ${i[0]} 不支持保价或无降价`)
+            await price()
+            if (i != cookiesArr.length - 1) {
+                await $.wait(20000)
+            }
+        }
+    }
+    if (allMessage) {
+        if ($.isNode()) await notify.sendNotify(`${$.name}`, `${allMessage}`);
+    }
+})()
+    .catch((e) => {
+        $.log('', `❌ ${$.name}, 失败! 原因: ${e}!`, '')
+    })
+    .finally(() => {
+        $.done();
+    })
+
+async function price() {
+    let num = 0
+    do {
+        $.token = $.jab.getToken() || ''
+        if ($.token) {
+            await siteppM_skuOnceApply();
+        }
+        num++
+    } while (num < 3 && !$.token)
+    await showMsg()
+}
+
+async function siteppM_skuOnceApply() {
+    let body = {
+        sid: "",
+        type: "25",
+        forcebot: "",
+        token: $.token,
+        feSt: $.token ? "s" : "f"
+    }
+    return new Promise(async resolve => {
+        $.post(taskUrl("siteppM_skuOnceApply", body), async (err, resp, data) => {
+            try {
+                if (err) {
+                    console.log(JSON.stringify(err))
+                    console.log(`${$.name} siteppM_skuOnceApply API请求失败，请检查网路重试`);
                 } else {
-                    await jstoken()
-                    // 申请请求
-                    p5 = {
-                        'url': `https://api.m.jd.com/api?appid=siteppM&functionId=siteppM_proApply&forcebot=&t=${$.timestamp}`,
-                        'form': {
-                            'body': JSON.stringify({
-                                "orderId": i[1],
-                                "orderCategory": "Others",
-                                "skuId": i[2],
-                                "sid": "",
-                                "type": "3",
-                                "refundtype": "1",
-                                "forcebot": "",
-                                "token": $.token,
-                                "feSt": "s"
-                            })
+                    if (safeGet(data)) {
+                        data = JSON.parse(data)
+                        if (data.flag) {
+                            await $.wait(25 * 1000)
+                            await siteppM_appliedSuccAmount()
+                        } else {
+                            console.log(`保价失败：${data.responseMessage}`)
+                            message += `保价失败：${data.responseMessage}\n`
                         }
-                    }
-                    await $.curl(p5)
-                    if ($.source.proSkuApplyId) {
-                        // 申请结果
-                        p6 = {
-                            'url': `https://api.m.jd.com/api?appid=siteppM&functionId=siteppM_moreApplyResult&forcebot=&t=${$.timestamp}`,
-                            'form': `body={"proSkuApplyIds":"${$.source.proSkuApplyId[0]}","type":"3"}`
-                        }
-                        await $.curl(p6)
-                        console.log(`商品: ${i[0]} `, $.haskey($.source, 'applyResults.0.applyResultVo.failTypeStr'))
-                    } else {
-                        console.log(`商品: ${i[0]} ${$.source.errorMessage}`)
                     }
                 }
+            } catch (e) {
+                $.logErr(e, resp);
+            } finally {
+                resolve(data);
             }
-        }
-    } catch (e) {}
+        })
+    })
 }
-async function extra() {
-    $.dom.window.close()
+function siteppM_appliedSuccAmount() {
+    let body = {
+        sid: "",
+        type: "25",
+        forcebot: "",
+        num: 15
+    }
+    return new Promise(resolve => {
+        $.post(taskUrl("siteppM_appliedSuccAmount", body), (err, resp, data) => {
+            try {
+                if (err) {
+                    console.log(JSON.stringify(err))
+                    console.log(`${$.name} siteppM_appliedSuccAmount API请求失败，请检查网路重试`)
+                } else {
+                    if (safeGet(data)) {
+                        data = JSON.parse(data)
+                        if (data.flag) {
+                            console.log(`保价成功：返还${data.succAmount}元`)
+                            message += `保价成功：返还${data.succAmount}元\n`
+                        } else {
+                            console.log(`保价失败：没有可保价的订单`)
+                        }
+                    }
+                }
+            } catch (e) {
+                $.logErr(e, resp)
+            } finally {
+                resolve(data)
+            }
+        })
+    })
 }
+
 async function jstoken() {
-    let {
-        JSDOM
-    } = jsdom;
+    const { JSDOM } = jsdom;
     let resourceLoader = new jsdom.ResourceLoader({
         userAgent: 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10.15; rv:91.0) Gecko/20100101 Firefox/91.0',
-        referrer: "https://msitepp-fm.jd.com/rest/priceprophone/priceProPhoneMenu",
+        referrer: "https://msitepp-fm.jd.com/rest/priceprophone/priceProPhoneMenu"
     });
     let virtualConsole = new jsdom.VirtualConsole();
-    var options = {
+    let options = {
         url: "https://msitepp-fm.jd.com/rest/priceprophone/priceProPhoneMenu",
         referrer: "https://msitepp-fm.jd.com/rest/priceprophone/priceProPhoneMenu",
         userAgent: 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10.15; rv:91.0) Gecko/20100101 Firefox/91.0',
         runScripts: "dangerously",
         resources: resourceLoader,
-        //  cookieJar,
         includeNodeLocations: true,
         storageQuota: 10000000,
         pretendToBeVisual: true,
         virtualConsole
     };
-    $.dom = new JSDOM(`<body><script src="https://js-nocaptcha.jd.com/statics/js/main.min.js"></script></body>`, options);
+    // const { window } = new JSDOM(``, options);
+    // const jdPriceJs = await downloadUrl("https://js-nocaptcha.jd.com/statics/js/main.min.js")
+    const dom = new JSDOM(`<body><script src="https://js-nocaptcha.jd.com/statics/js/main.min.js"></script></body>`, options);
     await $.wait(1000)
     try {
-        feSt = 's'
-        jab = new $.dom.window.JAB({
+        // window.eval(jdPriceJs)
+        // window.HTMLCanvasElement.prototype.getContext = () => {
+        //   return {};
+        // };
+        $.jab = new dom.window.JAB({
             bizId: 'jdjiabao',
             initCaptcha: false
         })
-        $.token = jab.getToken()
     } catch (e) {}
 }
+
+function downloadUrl(url) {
+    return new Promise(resolve => {
+        const options = { url, "timeout": 10000 };
+        $.get(options, async (err, resp, data) => {
+            let res = null
+            try {
+                if (err) {
+                    console.log(`⚠️网络请求失败`);
+                } else {
+                    res = data;
+                }
+            } catch (e) {
+                $.logErr(e, resp)
+            } finally {
+                resolve(res);
+            }
+        })
+    })
+}
+
+function showMsg() {
+    return new Promise(resolve => {
+        if (message) {
+            allMessage += `【京东账号${$.index}】${$.nickName || $.UserName}\n${message}${$.index !== cookiesArr.length ? '\n\n' : '\n\n'}`;
+        }
+        $.msg($.name, '', `【京东账号${$.index}】${$.nickName || $.UserName}\n${message}`);
+        resolve()
+    })
+}
+
+function taskUrl(functionId, body) {
+    return {
+        url: `${JD_API_HOST}api?appid=siteppM&functionId=${functionId}&forcebot=&t=${Date.now()}`,
+        body: `body=${encodeURIComponent(JSON.stringify(body))}`,
+        headers: {
+            "Host": "api.m.jd.com",
+            "Accept": "application/json",
+            "Content-Type": "application/x-www-form-urlencoded",
+            "Origin": "https://msitepp-fm.jd.com",
+            "Accept-Language": "zh-CN,zh-Hans;q=0.9",
+            "User-Agent": $.isNode() ? (process.env.JD_USER_AGENT ? process.env.JD_USER_AGENT : (require('./USER_AGENTS').USER_AGENT)) : ($.getdata('JDUA') ? $.getdata('JDUA') : "jdapp;iPhone;9.4.4;14.3;network/4g;Mozilla/5.0 (iPhone; CPU iPhone OS 14_3 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Mobile/15E148;supportJDSHWK/1"),
+            "Referer": "https://msitepp-fm.jd.com/",
+            "Accept-Encoding": "gzip, deflate, br",
+            "Cookie": cookie
+        }
+    }
+}
+
+function TotalBean() {
+    return new Promise(resolve => {
+        const options = {
+            url: "https://me-api.jd.com/user_new/info/GetJDUserInfoUnion",
+            headers: {
+                "Host": "me-api.jd.com",
+                "Accept": "*/*",
+                "User-Agent": "ScriptableWidgetExtension/185 CFNetwork/1312 Darwin/21.0.0",
+                "Accept-Language": "zh-CN,zh-Hans;q=0.9",
+                "Accept-Encoding": "gzip, deflate, br",
+                "Cookie": cookie
+            }
+        }
+        $.get(options, (err, resp, data) => {
+            try {
+                if (err) {
+                    $.logErr(err)
+                } else {
+                    if (data) {
+                        data = JSON.parse(data);
+                        if (data['retcode'] === "1001") {
+                            $.isLogin = false; //cookie过期
+                            return;
+                        }
+                        if (data['retcode'] === "0" && data.data && data.data.hasOwnProperty("userInfo")) {
+                            $.nickName = data.data.userInfo.baseInfo.nickname;
+                        }
+                    } else {
+                        console.log('京东服务器返回空数据');
+                    }
+                }
+            } catch (e) {
+                $.logErr(e, resp)
+            } finally {
+                resolve()
+            }
+        })
+    })
+}
+function safeGet(data) {
+    try {
+        if (typeof JSON.parse(data) == "object") {
+            return true;
+        }
+    } catch (e) {
+        console.log(e);
+        console.log(`京东服务器访问数据为空，请检查自身设备网络情况`);
+        return false;
+    }
+}
+function jsonParse(str) {
+    if (typeof str == "string") {
+        try {
+            return JSON.parse(str);
+        } catch (e) {
+            console.log(e);
+            $.msg($.name, '', '请勿随意在BoxJs输入框修改内容\n建议通过脚本去获取cookie')
+            return [];
+        }
+    }
+}
+
+// prettier-ignore
+function Env(t,e){"undefined"!=typeof process&&JSON.stringify(process.env).indexOf("GITHUB")>-1&&process.exit(0);class s{constructor(t){this.env=t}send(t,e="GET"){t="string"==typeof t?{url:t}:t;let s=this.get;return"POST"===e&&(s=this.post),new Promise((e,i)=>{s.call(this,t,(t,s,r)=>{t?i(t):e(s)})})}get(t){return this.send.call(this.env,t)}post(t){return this.send.call(this.env,t,"POST")}}return new class{constructor(t,e){this.name=t,this.http=new s(this),this.data=null,this.dataFile="box.dat",this.logs=[],this.isMute=!1,this.isNeedRewrite=!1,this.logSeparator="\n",this.startTime=(new Date).getTime(),Object.assign(this,e),this.log("",`🔔${this.name}, 开始!`)}isNode(){return"undefined"!=typeof module&&!!module.exports}isQuanX(){return"undefined"!=typeof $task}isSurge(){return"undefined"!=typeof $httpClient&&"undefined"==typeof $loon}isLoon(){return"undefined"!=typeof $loon}toObj(t,e=null){try{return JSON.parse(t)}catch{return e}}toStr(t,e=null){try{return JSON.stringify(t)}catch{return e}}getjson(t,e){let s=e;const i=this.getdata(t);if(i)try{s=JSON.parse(this.getdata(t))}catch{}return s}setjson(t,e){try{return this.setdata(JSON.stringify(t),e)}catch{return!1}}getScript(t){return new Promise(e=>{this.get({url:t},(t,s,i)=>e(i))})}runScript(t,e){return new Promise(s=>{let i=this.getdata("@chavy_boxjs_userCfgs.httpapi");i=i?i.replace(/\n/g,"").trim():i;let r=this.getdata("@chavy_boxjs_userCfgs.httpapi_timeout");r=r?1*r:20,r=e&&e.timeout?e.timeout:r;const[o,h]=i.split("@"),n={url:`http://${h}/v1/scripting/evaluate`,body:{script_text:t,mock_type:"cron",timeout:r},headers:{"X-Key":o,Accept:"*/*"}};this.post(n,(t,e,i)=>s(i))}).catch(t=>this.logErr(t))}loaddata(){if(!this.isNode())return{};{this.fs=this.fs?this.fs:require("fs"),this.path=this.path?this.path:require("path");const t=this.path.resolve(this.dataFile),e=this.path.resolve(process.cwd(),this.dataFile),s=this.fs.existsSync(t),i=!s&&this.fs.existsSync(e);if(!s&&!i)return{};{const i=s?t:e;try{return JSON.parse(this.fs.readFileSync(i))}catch(t){return{}}}}}writedata(){if(this.isNode()){this.fs=this.fs?this.fs:require("fs"),this.path=this.path?this.path:require("path");const t=this.path.resolve(this.dataFile),e=this.path.resolve(process.cwd(),this.dataFile),s=this.fs.existsSync(t),i=!s&&this.fs.existsSync(e),r=JSON.stringify(this.data);s?this.fs.writeFileSync(t,r):i?this.fs.writeFileSync(e,r):this.fs.writeFileSync(t,r)}}lodash_get(t,e,s){const i=e.replace(/\[(\d+)\]/g,".$1").split(".");let r=t;for(const t of i)if(r=Object(r)[t],void 0===r)return s;return r}lodash_set(t,e,s){return Object(t)!==t?t:(Array.isArray(e)||(e=e.toString().match(/[^.[\]]+/g)||[]),e.slice(0,-1).reduce((t,s,i)=>Object(t[s])===t[s]?t[s]:t[s]=Math.abs(e[i+1])>>0==+e[i+1]?[]:{},t)[e[e.length-1]]=s,t)}getdata(t){let e=this.getval(t);if(/^@/.test(t)){const[,s,i]=/^@(.*?)\.(.*?)$/.exec(t),r=s?this.getval(s):"";if(r)try{const t=JSON.parse(r);e=t?this.lodash_get(t,i,""):e}catch(t){e=""}}return e}setdata(t,e){let s=!1;if(/^@/.test(e)){const[,i,r]=/^@(.*?)\.(.*?)$/.exec(e),o=this.getval(i),h=i?"null"===o?null:o||"{}":"{}";try{const e=JSON.parse(h);this.lodash_set(e,r,t),s=this.setval(JSON.stringify(e),i)}catch(e){const o={};this.lodash_set(o,r,t),s=this.setval(JSON.stringify(o),i)}}else s=this.setval(t,e);return s}getval(t){return this.isSurge()||this.isLoon()?$persistentStore.read(t):this.isQuanX()?$prefs.valueForKey(t):this.isNode()?(this.data=this.loaddata(),this.data[t]):this.data&&this.data[t]||null}setval(t,e){return this.isSurge()||this.isLoon()?$persistentStore.write(t,e):this.isQuanX()?$prefs.setValueForKey(t,e):this.isNode()?(this.data=this.loaddata(),this.data[e]=t,this.writedata(),!0):this.data&&this.data[e]||null}initGotEnv(t){this.got=this.got?this.got:require("got"),this.cktough=this.cktough?this.cktough:require("tough-cookie"),this.ckjar=this.ckjar?this.ckjar:new this.cktough.CookieJar,t&&(t.headers=t.headers?t.headers:{},void 0===t.headers.Cookie&&void 0===t.cookieJar&&(t.cookieJar=this.ckjar))}get(t,e=(()=>{})){t.headers&&(delete t.headers["Content-Type"],delete t.headers["Content-Length"]),this.isSurge()||this.isLoon()?(this.isSurge()&&this.isNeedRewrite&&(t.headers=t.headers||{},Object.assign(t.headers,{"X-Surge-Skip-Scripting":!1})),$httpClient.get(t,(t,s,i)=>{!t&&s&&(s.body=i,s.statusCode=s.status),e(t,s,i)})):this.isQuanX()?(this.isNeedRewrite&&(t.opts=t.opts||{},Object.assign(t.opts,{hints:!1})),$task.fetch(t).then(t=>{const{statusCode:s,statusCode:i,headers:r,body:o}=t;e(null,{status:s,statusCode:i,headers:r,body:o},o)},t=>e(t))):this.isNode()&&(this.initGotEnv(t),this.got(t).on("redirect",(t,e)=>{try{if(t.headers["set-cookie"]){const s=t.headers["set-cookie"].map(this.cktough.Cookie.parse).toString();s&&this.ckjar.setCookieSync(s,null),e.cookieJar=this.ckjar}}catch(t){this.logErr(t)}}).then(t=>{const{statusCode:s,statusCode:i,headers:r,body:o}=t;e(null,{status:s,statusCode:i,headers:r,body:o},o)},t=>{const{message:s,response:i}=t;e(s,i,i&&i.body)}))}post(t,e=(()=>{})){if(t.body&&t.headers&&!t.headers["Content-Type"]&&(t.headers["Content-Type"]="application/x-www-form-urlencoded"),t.headers&&delete t.headers["Content-Length"],this.isSurge()||this.isLoon())this.isSurge()&&this.isNeedRewrite&&(t.headers=t.headers||{},Object.assign(t.headers,{"X-Surge-Skip-Scripting":!1})),$httpClient.post(t,(t,s,i)=>{!t&&s&&(s.body=i,s.statusCode=s.status),e(t,s,i)});else if(this.isQuanX())t.method="POST",this.isNeedRewrite&&(t.opts=t.opts||{},Object.assign(t.opts,{hints:!1})),$task.fetch(t).then(t=>{const{statusCode:s,statusCode:i,headers:r,body:o}=t;e(null,{status:s,statusCode:i,headers:r,body:o},o)},t=>e(t));else if(this.isNode()){this.initGotEnv(t);const{url:s,...i}=t;this.got.post(s,i).then(t=>{const{statusCode:s,statusCode:i,headers:r,body:o}=t;e(null,{status:s,statusCode:i,headers:r,body:o},o)},t=>{const{message:s,response:i}=t;e(s,i,i&&i.body)})}}time(t,e=null){const s=e?new Date(e):new Date;let i={"M+":s.getMonth()+1,"d+":s.getDate(),"H+":s.getHours(),"m+":s.getMinutes(),"s+":s.getSeconds(),"q+":Math.floor((s.getMonth()+3)/3),S:s.getMilliseconds()};/(y+)/.test(t)&&(t=t.replace(RegExp.$1,(s.getFullYear()+"").substr(4-RegExp.$1.length)));for(let e in i)new RegExp("("+e+")").test(t)&&(t=t.replace(RegExp.$1,1==RegExp.$1.length?i[e]:("00"+i[e]).substr((""+i[e]).length)));return t}msg(e=t,s="",i="",r){const o=t=>{if(!t)return t;if("string"==typeof t)return this.isLoon()?t:this.isQuanX()?{"open-url":t}:this.isSurge()?{url:t}:void 0;if("object"==typeof t){if(this.isLoon()){let e=t.openUrl||t.url||t["open-url"],s=t.mediaUrl||t["media-url"];return{openUrl:e,mediaUrl:s}}if(this.isQuanX()){let e=t["open-url"]||t.url||t.openUrl,s=t["media-url"]||t.mediaUrl;return{"open-url":e,"media-url":s}}if(this.isSurge()){let e=t.url||t.openUrl||t["open-url"];return{url:e}}}};if(this.isMute||(this.isSurge()||this.isLoon()?$notification.post(e,s,i,o(r)):this.isQuanX()&&$notify(e,s,i,o(r))),!this.isMuteLog){let t=["","==============📣系统通知📣=============="];t.push(e),s&&t.push(s),i&&t.push(i),console.log(t.join("\n")),this.logs=this.logs.concat(t)}}log(...t){t.length>0&&(this.logs=[...this.logs,...t]),console.log(t.join(this.logSeparator))}logErr(t,e){const s=!this.isSurge()&&!this.isQuanX()&&!this.isLoon();s?this.log("",`❗️${this.name}, 错误!`,t.stack):this.log("",`❗️${this.name}, 错误!`,t)}wait(t){return new Promise(e=>setTimeout(e,t))}done(t={}){const e=(new Date).getTime(),s=(e-this.startTime)/1e3;this.log("",`🔔${this.name}, 结束! 🕛 ${s} 秒`),this.log(),(this.isSurge()||this.isQuanX()||this.isLoon())&&$done(t)}}(t,e)}
